+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+
+### 🐛 Fixed
+  * Super multi-level baseRoutes handled #10
+
 ## v2.0.1 - 2016-11-05
 
 ### Changed